requirements.txt
discord.py==1.7.3
python-decouple==3.5